Define additional 1.1 facets and move inquiry of allowed facets to a type validator helper to differentiate between 1.0 types vs. 1.1 types.
git-svn-id: https://svn.apache.org/repos/asf/xerces/java/branches/xml-schema-1.1-dev@885800 13f79535-47bb-0310-9956-ffa450edef68

+++ samples/simpletype/SimpleTypeUsage.java
@@ -191,7 +191,7 @@ public void querySimpleType(XSSimpleType simpleType){
 
     // 'facets' property
     // gives bit combination of the constants defined in XSSimpleType interface.
-    short facets = simpleType.getDefinedFacets() ;
+    int facets = simpleType.getDefinedFacets() ;
     printFacets(facets);
 
     //'final' property
@@ -270,7 +270,7 @@ void printCardinality (boolean isFinite){
 
 }//printCardinality()
 
-void printFacets(short facets){
+void printFacets(int facets){
 
     System.err.println("'facets' present \t\t: " );
